+++ server/animals.routes.js
@@ -12,15 +12,8 @@ const getAllAnimals = async (req, res) => {
   res.json(animals)
 }
 
-// const getRandomAnimal = async (req, res) => {
-//   const animalId = req.params.animalId
-//   const foundAnimal = await Animal.find(animalId)
-//   res.json(foundAnimal)
-// }
-
 const router = express.Router()
 
 router.get("/api/animals", getAllAnimals)
-//router.get('/api/animals/:animalId', getRandomAnimal)
 
 export default router // as AnimalsRoutes to server.js
